Make the beginning of a pratt parser
full list of changes:

- make a pratt parser which can parse ints, +-*/ and parens
- add support for parser errors
- moved stringify_token_type back to lexer.c as it's needed for error
messages (i plan not to compile with debug.c)
- add dedicated free_token function so that the parser can free tokens
as it goes instead of waiting until the end
- make the linked lists of tokens also keep track of the tail (for O(1)
push and pop)

+++ src/utils.c
@@ -19,7 +19,7 @@ char *read_file(char *path) {
 	fseek(file, 0, SEEK_SET);
 
 	// 0.5GB (probably don't want to load that into memory)
-	const int MAX_FILE_SIZE = 500000000;
+	const size_t MAX_FILE_SIZE = 500000000;
 
 	if (file_length > MAX_FILE_SIZE) {
 		printf("Error: file must be less than 0.5GB\n");
@@ -96,7 +96,7 @@ void append_str_and_free(char **dest, char *src) {
 }
 
 char *alloc_num_as_str(size_t number) {
-	int string_len;
+	size_t string_len;
 	if (number == 0) string_len = 2;
 	else string_len = (int)log10(number) + 2;
